Fixes #3 - add delete for multiple therapists and confirm dialog (also for single delete #2)

+++ tl-webapp/src/app/therapists/overview/card/card.component.ts
@@ -1,13 +1,14 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { MatCard, MatCardContent, MatCardHeader, MatCardSubtitle, MatCardTitle } from '@angular/material/card';
-import { RequestChip, TherapistCardTs, TherapistService } from 'shared';
+import { ConfirmDialog, RequestChip, TherapistCardTs, TherapistService } from 'shared';
 import { NgIf } from '@angular/common';
 import { BusinessHoursList } from '../../details/business-hours-list/business-hours-list';
-import { CARD_PATH } from '../../../translation-paths';
+import { CARD_PATH, OVERVIEW_PATH } from '../../../translation-paths';
 import { TranslocoPipe } from '@ngneat/transloco';
 import { MatIconButton } from '@angular/material/button';
 import { MatIcon } from '@angular/material/icon';
 import { MatMenu, MatMenuItem, MatMenuTrigger } from '@angular/material/menu';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-card',
@@ -38,18 +39,31 @@ export class CardComponent {
   @Output()
   updated = new EventEmitter<void>();
 
-  constructor(private therapistService: TherapistService) {
+  constructor(private therapistService: TherapistService, private dialog: MatDialog) {
   }
 
-  onDeleteClicked($event: Event) {
-    $event.stopPropagation();
-    if (!this.therapist?.id) {
-      return;
-    }
-    this.therapistService.deleteTherapist(this.therapist.id.toString())
-      .then(() => {
-        this.updated.emit();
-      });
+  onDeleteClicked() {
+    const dialogRef = this.dialog.open(ConfirmDialog, {
+      data: {
+        titlePath: OVERVIEW_PATH + ".deleteConfirm.title",
+        messagePath: CARD_PATH + ".deleteConfirm.message",
+        messageParams: {name: this.therapist?.name},
+        confirmButtonPath: OVERVIEW_PATH + ".deleteConfirm.confirm",
+        warn: true,
+      }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result === true) {
+        if (!this.therapist?.id) {
+          return;
+        }
+        this.therapistService.deleteTherapist(this.therapist.id.toString())
+          .then(() => {
+            this.updated.emit();
+          });
+      }
+    });
   }
 
 }
